requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dateutils==0.6.6
decorator==4.4.0
defusedxml==0.6.0
Django==2.1
django-environ==0.4.5
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
graphviz==0.12
grpcio==1.23.0
h5py==2.9.0
idna==2.8
intel-openmp==2019.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl==2019.0
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.0
nsepy==0.7
nsetools==1.0.10
numpy==1.16.4
opencv-python==4.1.1.26
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psycopg2==2.8.3
ptyprocess==0.6.0
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.4.2
PyJWT==1.7.1
pymc3==3.7
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
statsmodels==0.10.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
tornado==6.0.3
tqdm==4.36.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
